feat: rename visits to harvest and unify date formatting
