+++ src/components/history/HarvestList.tsx
@@ -3,15 +3,7 @@ import { Button } from "@/components/ui/button";
 import type { VisitHistory } from "@/types";
 import { useT } from "@/i18n";
 import { Stars } from "@/components/common/Stars";
-
-function formatDate(str: string) {
-  const d = new Date(str);
-  if (Number.isNaN(d.getTime())) return str;
-  const dd = String(d.getDate()).padStart(2, "0");
-  const mm = String(d.getMonth() + 1).padStart(2, "0");
-  const yyyy = d.getFullYear();
-  return `${dd}/${mm}/${yyyy}`;
-}
+import { formatDate } from "@/utils";
 
 interface HarvestListProps {
   items: VisitHistory[];
