+++ src/i18n/common.ts
@@ -35,7 +35,7 @@ export default {
   "Aucun coin enregistré.": { fr: "Aucun coin enregistré.", en: "No spot saved." },
   "modifier": { fr: "modifier", en: "edit" },
   "Espèces :": { fr: "Espèces :", en: "Species:" },
-  "Dernière visite :": { fr: "Dernière visite :", en: "Last visit:" },
+  "Dernière cueillette :": { fr: "Dernière cueillette :", en: "Last harvest:" },
   "Itinéraire": { fr: "Itinéraire", en: "Route" },
   "Données stockées localement. Accès hors‑ligne.": {
     fr: "Données stockées localement. Accès hors‑ligne.",
@@ -60,7 +60,7 @@ export default {
   "Supprimer": { fr: "Supprimer", en: "Delete" },
   "Supprimer ce coin ?": { fr: "Supprimer ce coin ?", en: "Delete this spot?" },
   "Ajouter un champignon…": { fr: "Ajouter un champignon…", en: "Add a mushroom…" },
-  "Dernière visite": { fr: "Dernière visite", en: "Last visit" },
+  "Dernière cueillette": { fr: "Dernière cueillette", en: "Last harvest" },
   "Photos": { fr: "Photos", en: "Photos" },
   "Importer des photos": { fr: "Importer des photos", en: "Import photos" },
   "Annuler": { fr: "Annuler", en: "Cancel" },
@@ -96,10 +96,10 @@ export default {
     fr: "© OpenStreetMap contributors | MapLibre",
     en: "© OpenStreetMap contributors | MapLibre",
   },
-  "Visites": { fr: "Visites", en: "Visits" },
-  "Visite": { fr: "Visite", en: "Visit" },
+  "Cueillettes": { fr: "Cueillettes", en: "Harvests" },
+  "Cueillette du {date}": { fr: "Cueillette du {date}", en: "Harvest on {date}" },
   "Ajouter une cueillette": { fr: "Ajouter une cueillette", en: "Add a harvest" },
-  "Aucune visite enregistrée.": { fr: "Aucune visite enregistrée.", en: "No visit recorded." },
+  "Aucune cueillette enregistrée.": { fr: "Aucune cueillette enregistrée.", en: "No harvest recorded." },
   "Note:": { fr: "Note:", en: "Rating:" },
   "Galerie": { fr: "Galerie", en: "Gallery" },
   "Photo de la cueillette {n}": { fr: "Photo de la cueillette {n}", en: "Photo of harvest {n}" },
